feat: Request address and sign message for account creation
feat: Implement Xverse signing for Hive account creation

feat: Add create account functionality

feat: Add "Create Account" button and modal

Changed the transaction IDs in the contracts.js, server.js, ui.js, and wallet.js files to use the updated 'spkccT_' prefix.

Updated the API endpoint URLs in the contracts.js, login.js, and ui.js files.

Added console log for signature response

Added the Bitcoin signature to the metadata.

Added the Bitcoin signature and message to the Hive profile metadata.

Used the same method for calling the Hive API as used to get the profile picture and used the Hive Keychain to update the profile metadata.

Fetched Hive profile JSON metadata, updated it with Bitcoin address and signature, and broadcasted the updated profile to the Hive blockchain.

Implemented a process to first obtain the user's Bitcoin address before signing the message.

Added event listener and functionality to link BTC wallet using Xverse.

Replaced profile button with link BTC wallet button.

Added a dropdown menu to the profile picture in the nav bar, which is only displayed when the user is logged in. The dropdown menu contains "Profile" and "Logout" options.

Implemented a dropdown menu for the profile picture in the nav bar with "Profile" and "Logout" options when logged in.

Extracted filename and extension during file upload and improved login status messaging.